Enviroment_Section_update
I was able to edit/reorient objects and add directional light in Unity and saved changes. Looking at the scale, it needs to be readjusted to a smaller scale, the plane is too big for just a room. That could be done in Unity, but it will have to be done in every object by separate.