Enviroment_sections_update
I was able to export section 2 and placed next to section 1 to give it a some sort of connection-look. Section 1 had missing all stones while exporting section 2. For some reason the stones were in the hierarchy but not showing. I deleted empty/invisible stones in the hierarchy and cloned stones from section 2 and added them to section 1 and saved changes. Section 2 has a missing animation in the fountain. most probably I didn't exported it with the other assets.