Environment_update
I added 2 sections more to the environment, I realized I'm missing the animation on the fountain and waterfall. Section 4 had some scripts embedded I'll had to remove it and update it again. Connecting the sections I realized there are two missing walls for section 2, 3 and 4 that must be added. There are white line showing on the terrain model in the game view. I tried to fix it with a different shader, it did not make any difference except for the light. I'll review the model and update again.