Enviroment_update
All terrains have been added to the Theseus file on Unity as prefabs. I had some challenges with the GUID that made some assets not visible in the scene. I fixed it by adding the missing assets again. There are 8 terrains in total. that includes the entrance of the maze that would be the las part of the game. I added a secondary folder "modifiedPrefab" with the saved changes I made on each prefab, 8 in total.